Added integer type massaging to NewNode
- Removed int64 casts that are now handled by integer type massaging
- Added tests for node.go
- Restructured parser_test.go and improved coverage

[#69754064]

+++ compare/compare.go
@@ -153,7 +153,7 @@ func jobMap(jobs []yaml.Node) map[string]yaml.Node {
 
 	for index, job := range jobs {
 		attrs, ok := job.Value().(map[string]yaml.Node)
-		attrs["index"] = yaml.NewNode(int64(index), job.SourceName())
+		attrs["index"] = yaml.NewNode(index, job.SourceName())
 
 		name, ok := yaml.FindString(job, "name")
 		if !ok {
